agents/groq_llama_vision_agent/requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
Pillow
python-base64